refactor: remove preset card
