+++ frontend/src/component/feature/FeatureStrategy/FeatureStrategyEmpty/AddFromTemplateCard/AddFromTemplateCard.tsx
@@ -1,4 +1,5 @@
 import { ElementType, FC } from 'react';
+import { Card, CardContent, Typography, styled, Box } from '@mui/material';
 import { ChangeRequestDialogue } from 'component/changeRequest/ChangeRequestConfirmDialog/ChangeRequestConfirmDialog';
 import useFeatureStrategyApi from 'hooks/api/actions/useFeatureStrategyApi/useFeatureStrategyApi';
 import useToast from 'hooks/useToast';
@@ -6,7 +7,8 @@ import { AddStrategyMessage } from 'component/changeRequest/ChangeRequestConfirm
 import { useChangeRequestsEnabled } from 'hooks/useChangeRequestsEnabled';
 import { useChangeRequestAddStrategy } from 'hooks/useChangeRequestAddStrategy';
 import { formatUnknownError } from 'utils/formatUnknownError';
-import { PresetCard } from './PresetCard/PresetCard';
+import PermissionButton from 'component/common/PermissionButton/PermissionButton';
+import { CREATE_FEATURE_STRATEGY } from 'component/providers/AccessProvider/permissions';
 import { IFeatureStrategyPayload } from 'interfaces/strategy';
 
 interface IAddFromTemplateCardProps {
@@ -18,6 +20,12 @@ interface IAddFromTemplateCardProps {
     Icon: ElementType;
     onAfterAddStrategy: () => void;
 }
+
+const StyledCard = styled(Card)(({ theme }) => ({
+    display: 'flex',
+    flexDirection: 'column',
+    borderRadius: theme.shape.borderRadiusMedium,
+}));
 
 export const AddFromTemplateCard: FC<IAddFromTemplateCardProps> = ({
     title,
@@ -41,7 +49,7 @@ export const AddFromTemplateCard: FC<IAddFromTemplateCardProps> = ({
         onChangeRequestAddStrategyClose,
     } = useChangeRequestAddStrategy(projectId, featureId, 'addStrategy');
 
-    const onStrategy = async () => {
+    const onStrategyAdd = async () => {
         try {
             if (isChangeRequestEnabled) {
                 onChangeRequestAddStrategy(environmentId, strategy);
@@ -61,15 +69,50 @@ export const AddFromTemplateCard: FC<IAddFromTemplateCardProps> = ({
 
     return (
         <>
-            <PresetCard
-                title={title}
-                Icon={Icon}
-                onClick={onStrategy}
-                projectId={projectId}
-                environmentId={environmentId}
-            >
-                {children}
-            </PresetCard>
+            <StyledCard variant="outlined">
+                <CardContent
+                    sx={{
+                        display: 'flex',
+                        flexDirection: 'column',
+                        flexGrow: 1,
+                    }}
+                >
+                    <Typography
+                        variant="body1"
+                        fontWeight="medium"
+                        sx={{ mb: 0.5, display: 'flex', alignItems: 'center' }}
+                    >
+                        <Icon color="disabled" sx={{ mr: 1 }} /> {title}
+                    </Typography>
+                    <Typography
+                        variant="body2"
+                        color="text.secondary"
+                        component="p"
+                    >
+                        {children}
+                    </Typography>
+
+                    <Box
+                        sx={{
+                            ml: 'auto',
+                            mt: 'auto',
+                            pt: 1,
+                            mr: { xs: 'auto', sm: 0 },
+                        }}
+                    >
+                        <PermissionButton
+                            permission={CREATE_FEATURE_STRATEGY}
+                            projectId={projectId}
+                            environmentId={environmentId}
+                            variant="outlined"
+                            size="small"
+                            onClick={onStrategyAdd}
+                        >
+                            Use template
+                        </PermissionButton>
+                    </Box>
+                </CardContent>
+            </StyledCard>
             <ChangeRequestDialogue
                 isOpen={changeRequestDialogDetails.isOpen}
                 onClose={onChangeRequestAddStrategyClose}
